Rename project to MiHealth AmapFix
