+++ app/build.gradle
@@ -8,11 +8,11 @@ kotlin {
 }
 
 android {
-    namespace "io.github.mihealthgaodefix"
+    namespace "io.github.mihealthamapfix"
     compileSdk 34
 
     defaultConfig {
-        applicationId "io.github.mihealthgaodefix"
+        applicationId "io.github.mihealthamapfix"
         minSdk 24
         targetSdk 34
         versionCode 2
