+++ app/proguard-rules.pro
@@ -1,7 +1,7 @@
 # 让 LSPosed 能通过类名加载到入口与 Hooker（名字不能被混淆/删除）
--keep class io.github.amapnotifyfix.ModernEntry { *; }
--keep class io.github.amapnotifyfix.ModernEntry$IsMipmapHooker { *; }
--keep class io.github.amapnotifyfix.LegacyInit { *; }
+-keep class io.github.mihealthamapfix.ModernEntry { *; }
+-keep class io.github.mihealthamapfix.ModernEntry$IsMipmapHooker { *; }
+-keep class io.github.mihealthamapfix.LegacyInit { *; }
 
 # 保留 libxposed Hooker 的静态回调方法签名（before/after）
 -keepclassmembers class * implements io.github.libxposed.api.XposedInterface$Hooker {
