+++ app/src/main/AndroidManifest.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="io.github.mihealthgaodefix">
+    package="io.github.mihealthamapfix">
 
     <application
-        android:label="Mi Fitness Amap Nav Fix"
+        android:label="MiHealth AmapFix"
         android:description="@string/module_description"
         android:allowBackup="false"
         android:supportsRtl="true">
