+++ app/src/main/assets/xposed_init
@@ -1,1 +1,1 @@
-io.github.amapnotifyfix.LegacyInit
+io.github.mihealthamapfix.LegacyInit
