+++ app/src/main/resources/META-INF/xposed/java_init.list
@@ -1,1 +1,1 @@
-io.github.amapnotifyfix.ModernEntry
+io.github.mihealthamapfix.ModernEntry
